+++ src/app/chits/chits.component.ts
@@ -12,7 +12,11 @@ import { NgForm } from '@angular/forms';
 export class ChitsComponent implements OnInit {
 
   
-  url:string="chits"
+  url:string="chits";
+  noOfChits=36;
+  noOfChitsPaid=0;
+  totalChitAmount:number=180000;
+  totalAmountPaid:number=0;
 
   constructor(private apiService:ApiService) { 
     this.get();
@@ -37,15 +41,19 @@ export class ChitsComponent implements OnInit {
   }
 
   get(){
-    this.apiService.getObjects(this.url).then(result=>{
+    this.totalAmountPaid=0;
+    this.noOfChitsPaid=0;
+    this.apiService.getOrderedObjects(this.url,"updatedOn","desc").then(result=>{
       this.chits=(result.map(x=>{
+        this.totalAmountPaid+=x.payload.doc.data()['actualAmount'];
+        this.noOfChitsPaid++;
         return {
           id:x.payload.doc.id,
           month:x.payload.doc.data()['month'],
           year:x.payload.doc.data()['year'],
           actualAmount:x.payload.doc.data()['actualAmount'],
           paidAmount:x.payload.doc.data()['paidAmount'],
-          profit:x.payload.doc.data()['profit'],
+          profit:x.payload.doc.data()['profit']
         }
       }));
      
@@ -62,8 +70,10 @@ export class ChitsComponent implements OnInit {
   }
 
   delete(id:string){
-    this.apiService.deleteObject(id,this.url);
-    this.get();
+    if (confirm("Are you sure you want to delete?")) {
+      this.apiService.deleteObject(id,this.url);
+      this.get();
+    }
   }
 
 }
